# docs/swagger/swagger.json
{
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "基于Sherpa-ONNX的语音识别(STT)和语音合成(TTS)服务API文档",
        "title": "AeroSpeech-ONNX API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/zhangjun/AeroSpeech-ONNX",
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/stt/batch": {
            "post": {
                "description": "批量识别多个音频文件",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "STT"
                ],
                "summary": "批量识别",
                "parameters": [
                    {
                        "description": "批量识别请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.BatchRecognizeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "识别成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/stt/recognize": {
            "post": {
                "description": "上传音频文件进行语音识别",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "STT"
                ],
                "summary": "文件上传识别",
                "parameters": [
                    {
                        "type": "file",
                        "description": "音频文件",
                        "name": "audio",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "识别成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/tts/batch": {
            "post": {
                "description": "批量将多个文本合成为语音音频",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TTS"
                ],
                "summary": "批量合成",
                "parameters": [
                    {
                        "description": "批量合成请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.BatchSynthesizeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "合成成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/tts/synthesize": {
            "post": {
                "description": "将文本合成为语音音频",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "audio/wav"
                ],
                "tags": [
                    "TTS"
                ],
                "summary": "文本合成",
                "parameters": [
                    {
                        "description": "合成请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.SynthesizeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "音频文件",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.BatchRecognizeRequest": {
            "type": "object",
            "required": [
                "files"
            ],
            "properties": {
                "files": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "handlers.BatchSynthesizeRequest": {
            "type": "object",
            "required": [
                "texts"
            ],
            "properties": {
                "texts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.SynthesizeRequest"
                    }
                }
            }
        },
        "handlers.SynthesizeRequest": {
            "type": "object",
            "required": [
                "text"
            ],
            "properties": {
                "speaker_id": {
                    "type": "integer"
                },
                "speed": {
                    "type": "number"
                },
                "text": {
                    "type": "string"
                }
            }
        }
    }
}